pyproject.toml
[tool.poetry]
name = "personal-networking"
version = "0.1.0"
description = ""
authors = ["Dmitrii Sychev"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
uvicorn = "^0.40.0"
pydantic-settings = "^2.12.0"
pydantic = "^2.12.5"
fastapi = "^0.128.0"
loguru = "^0.7.3"
sqlalchemy = "^2.0.45"
httpx = "^0.28.1"
python-jose = {version = "^3.5.0", extras = ["cryptography"]}
asyncpg = "^0.31.0"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

